moved general dashboard out of testing folder removed unused files - Sandbox and GeneralTable
