+++ RCHIWebDash/RCHIWebDashBoard/src/General Dashboard/constants.js
@@ -1,0 +1,152 @@
+const GEN_SUBPOP_ORDER = [
+    'Individuals',
+    'Adults (≥25)',
+    'Youth (18-24)',
+    'Children (≤17)',
+    'Unknown Ages',
+    'Households (Interview)',
+    'Families w/ Children (Interview)',
+    'Chronically Homeless',
+    'Substance Abuse',
+    'Mental Health Conditions',
+    'Veterans',
+    'Pet Owners',
+]
+
+
+
+const SHELTERED_SUBPOP_ORDER = [
+    'Individuals',
+    'Adults (≥25)',
+    'Youth (18-24)',
+    'Children (≤17)',
+    'Unknown Ages',
+    'Male',
+    'Female',
+    'Transgender',
+    'Gender Non-Con',
+    'Unknown Gender',
+    'Hispanic',
+    'NonHispanic',
+    'Unknown Ethnicity',
+    'Asian',
+    'American Indian',
+    'Black',
+    'White',
+    'Multiple Races',
+    'Native Hawaiian',
+    'Unknown Race',
+    'Chronically Homeless',
+    'Substance Abuse',
+    'Veterans',
+    'Pet Owners',
+  ];
+
+
+  const FILTERED_COUNTS = [
+    'Total',
+    'Veteran No',
+    'Unknown Veteran',
+    'Not Chronically Homeless',
+    'No Substance Abuse',
+    'Unknown Substance Abuse',
+    'PTSD',
+    'No PTSD',
+    'Unknown PTSD',
+    'Mental Health Conditions',
+    'No Mental Health Conditions',
+    'Unknown Mental Health Conditions',
+    'Physical Disability',
+    'No Physical Disability',
+    'Unknown Physical Disability',
+    'Developmental Disability',
+    'No Developmental Disability',
+    'Unknown Developmental Disability',
+    'Brain Injury',
+    'No Brain Injury',
+    'Unknown Brain Injury',
+    'Victim of Domestic Violence',
+    'Not Victim of Domestic Violence',
+    'Unknown Victim of Domestic Violence',
+    'AIDS or HIV',
+    'No AIDS or HIV',
+    'Unknown AIDS or HIV',
+    'Jail Release 90 Days: Probation',
+    'Jail Release 90 Days: Parole',
+    'Jail Release 90 Days: Completed Sentence',
+    'Jail Release 90 Days: (Unspecified)',
+    'Jail Release 12 Months: Probation',
+    'Jail Release 12 Months: Parole',
+    'Jail Release 12 Months: Completed Sentence',
+    'Jail Release 12 Months: (Unspecified)',
+    'No Jail',
+    'Unknown Jail',
+    'Woods',
+    'Vehicle',
+    'UnderBridge',
+    'Street',
+    'Park',
+    'Other',
+    'Bus',
+    'Abandoned Building',
+    'Adults Only',
+    'Children Only',
+    'Adults and Children',
+    'Families w/ Children',
+    'Not Veteran',
+    'Households',
+    'Encampment',
+    'Couch',
+    'Households (Interview)',
+    'Families w/ Children (Interview)'
+  
+  ]
+  const FILTER_COLUMNS = [
+    'Total',
+    'Not Veteran',
+    'Veteran No',
+    'Unknown Veteran',
+    'No Substance Abuse',
+    'Unknown Substance Abuse',
+    'PTSD',
+    'No PTSD',
+    'Unknown PTSD',
+    'No Mental Health Conditions',
+    'Unknown Mental Health Conditions',
+    'Physical Disability',
+    'No Physical Disability',
+    'Unknown Physical Disability',
+    'Developmental Disability',
+    'No Developmental Disability',
+    'Unknown Developmental Disability',
+    'Brain Injury',
+    'No Brain Injury',
+    'Unknown Brain Injury',
+    'Victim of Domestic Violence',
+    'Not Victim of Domestic Violence',
+    'Unknown Victim of Domestic Violence',
+    'AIDS or HIV',
+    'No AIDS or HIV',
+    'Unknown AIDS or HIV',
+    'Jail Release 90 Days: Probation',
+    'Jail Release 90 Days: Parole',
+    'Jail Release 90 Days: Completed Sentence',
+    'Jail Release 90 Days: (Unspecified)',
+    'Jail Release 12 Months: Probation',
+    'Jail Release 12 Months: Parole',
+    'Jail Release 12 Months: Completed Sentence',
+    'Jail Release 12 Months: (Unspecified)',
+    'No Jail',
+    'Unknown Jail',
+    'Chronically Homeless	447',
+    'Not Chronically Homeless',
+    'Couch'
+  ];
+
+
+  export {
+    GEN_SUBPOP_ORDER,
+    SHELTERED_SUBPOP_ORDER,
+    FILTERED_COUNTS,
+    FILTER_COLUMNS
+  }
