+++ RCHIWebDash/RCHIWebDashBoard/src/components/Content.js
@@ -2,7 +2,6 @@ import React, { Component } from "react";
 import { Switch, Route } from "react-router-dom";
 import { router } from "../components/Utilities/constants/routing";
 
-import GeneralDash from "./DemoDashboards/GeneralDash";
 import VeteranDash from "./DemoDashboards/VeteranDash";
 import EditableCharts from "./EditableChartsDash/EditableCharts";
 import SupervisorialDistricts from "./SupervisoralDistrictDash/SupervisorialDistricts";
@@ -15,13 +14,11 @@ import SeniorsDashboard from "./../SeniorsDashBoard/SeniorsDashBoard";
 import UnshelteredSubpopulationTrends from "../Unsheltered Trends/UnshelteredSubpopulationTrends";
 import UnsheleteredSupervisoryDistricts from "../Unsheltered Supervisory District/UnshelteredSupervisoryDistricts";
 import LLMap from "./map/map";
-import GeneralDashboard from "../components/TestingBranch/GeneralDashboard";
+import GeneralDashboard from "../General Dashboard/GeneralDashboard";
 
 import APITest from "./APITests/ApiTest";
 
-import Sandbox from "./TestingBranch/Sandbox";
 import GeneralPieChart from "./charts/PieChart";
-import GeneralTable from "../Unsheltered Sheltered General/GeneralTable";
 import CityTable from "../Unsheltered Cities/CityTable";
 import errorPage from "./404Page";
 
@@ -57,7 +54,7 @@ export default class Content extends Component {
                 exact
                 path="/"
                 component={() => (
-                  <GeneralTable router={router} Tables={this.state.Tables} />
+                  <GeneralDashboard router={router} Tables={this.state.Tables} />
                 )}
               />
               
@@ -65,7 +62,7 @@ export default class Content extends Component {
                 exact
                 path="/GeneralTable"
                 component={() => (
-                  <GeneralTable router={router} Tables={this.state.Tables} />
+                  <GeneralDashboard router={router} Tables={this.state.Tables} />
                 )}
               />
               <Route
@@ -142,13 +139,7 @@ export default class Content extends Component {
                   <VeteranDash router={router} Tables={this.state.Tables} />
                 )}
               />
-              <Route
-                exact
-                path="/GeneralDash"
-                component={() => (
-                  <GeneralDash router={router} Tables={this.state.Tables} />
-                )}
-              />
+              
               <Route
                 exact
                 path="/TabChartDash"
@@ -191,24 +182,9 @@ export default class Content extends Component {
                   <APITest router={router} Tables={this.state.Tables} />
                 )}
               />
-              <Route
-                exact
-                path="/generalDashboard"
-                component={() => (
-                  <GeneralDashboard
-                    router={router}
-                    Tables={this.state.Tables}
-                  />
-                )}
-              />
+             
             
-              <Route
-                exact
-                path="/Sandbox"
-                component={() => (
-                  <Sandbox router={router} Tables={this.state.Tables} />
-                )}
-              />
+             
               <Route component={errorPage} />
             </Switch>
           </div>
